Moves common repository code to base package.
