+++ internal/repositories/aftermarket/aftermarket_devices.go
@@ -11,7 +11,7 @@ import (
 	"github.com/DIMO-Network/identity-api/graph/model"
 	gmodel "github.com/DIMO-Network/identity-api/graph/model"
 	"github.com/DIMO-Network/identity-api/internal/helpers"
-	"github.com/DIMO-Network/identity-api/internal/repositories"
+	"github.com/DIMO-Network/identity-api/internal/repositories/base"
 	"github.com/DIMO-Network/identity-api/models"
 	"github.com/DIMO-Network/mnemonic"
 	"github.com/ethereum/go-ethereum/common"
@@ -23,7 +23,7 @@ import (
 )
 
 type Repository struct {
-	*repositories.Repository
+	*base.Repository
 }
 
 // GetOwnedAftermarketDevices godoc
@@ -34,7 +34,7 @@ type Repository struct {
 // @Param last [*int] "the number of devices to return from previous pages"
 // @Param before [*string] "base64 string representing a device tokenID. Pointer to where we start fetching devices from previous pages"
 func (r *Repository) GetAftermarketDevices(ctx context.Context, first *int, after *string, last *int, before *string, filterBy *gmodel.AftermarketDevicesFilter) (*gmodel.AftermarketDeviceConnection, error) {
-	limit, err := helpers.ValidateFirstLast(first, last, repositories.MaxPageSize)
+	limit, err := helpers.ValidateFirstLast(first, last, base.MaxPageSize)
 	if err != nil {
 		return nil, err
 	}
@@ -146,7 +146,7 @@ func (r *Repository) GetAftermarketDevices(ctx context.Context, first *int, afte
 }
 
 func (r *Repository) GetAftermarketDevice(ctx context.Context, by gmodel.AftermarketDeviceBy) (*gmodel.AftermarketDevice, error) {
-	if repositories.CountTrue(by.Address != nil, by.TokenID != nil, by.Serial != nil) != 1 {
+	if base.CountTrue(by.Address != nil, by.TokenID != nil, by.Serial != nil) != 1 {
 		return nil, gqlerror.Errorf("Pass in exactly one of `address`, `id`, or `serial`.")
 	}
