+++ internal/repositories/dcn/dcn.go
@@ -8,7 +8,7 @@ import (
 
 	gmodel "github.com/DIMO-Network/identity-api/graph/model"
 	"github.com/DIMO-Network/identity-api/internal/helpers"
-	"github.com/DIMO-Network/identity-api/internal/repositories"
+	"github.com/DIMO-Network/identity-api/internal/repositories/base"
 	"github.com/DIMO-Network/identity-api/models"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/vektah/gqlparser/v2/gqlerror"
@@ -18,7 +18,7 @@ import (
 )
 
 type Repository struct {
-	*repositories.Repository
+	*base.Repository
 }
 
 type DCNCursor struct {
@@ -39,7 +39,7 @@ func DCNToAPI(d *models.DCN) *gmodel.Dcn {
 }
 
 func (r *Repository) GetDCN(ctx context.Context, by gmodel.DCNBy) (*gmodel.Dcn, error) {
-	if repositories.CountTrue(len(by.Node) != 0, by.Name != nil) != 1 {
+	if base.CountTrue(len(by.Node) != 0, by.Name != nil) != 1 {
 		return nil, gqlerror.Errorf("Provide exactly one of `name` or `node`.")
 	}
 
@@ -79,7 +79,7 @@ func (r *Repository) GetDCNByName(ctx context.Context, name string) (*gmodel.Dcn
 var dcnCursorColumnsTuple = "(" + models.DCNColumns.MintedAt + ", " + models.DCNColumns.Node + ")"
 
 func (r *Repository) GetDCNs(ctx context.Context, first *int, after *string, last *int, before *string, filterBy *gmodel.DCNFilter) (*gmodel.DCNConnection, error) {
-	limit, err := helpers.ValidateFirstLast(first, last, repositories.MaxPageSize)
+	limit, err := helpers.ValidateFirstLast(first, last, base.MaxPageSize)
 	if err != nil {
 		return nil, err
 	}
