+++ internal/repositories/dcn/dcn_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/DIMO-Network/identity-api/internal/config"
 	"github.com/DIMO-Network/identity-api/internal/helpers"
 	test "github.com/DIMO-Network/identity-api/internal/helpers"
-	"github.com/DIMO-Network/identity-api/internal/repositories"
+	"github.com/DIMO-Network/identity-api/internal/repositories/base"
 	"github.com/DIMO-Network/identity-api/models"
 	"github.com/DIMO-Network/shared/db"
 	"github.com/ethereum/go-ethereum/common"
@@ -37,7 +37,7 @@ func (o *DCNRepoTestSuite) SetupSuite() {
 		DIMORegistryAddr:    "0x2daeF6FF0E2B61aaBF9ADDE1aFDAfDb4f0f1E660",
 		DIMORegistryChainID: 80001,
 	}
-	o.repo = &Repository{repositories.New(o.pdb, o.settings)}
+	o.repo = &Repository{base.New(o.pdb, o.settings)}
 }
 
 // TearDownTest after each test truncate tables
