+++ internal/repositories/manufacturer/manufacturer_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/DIMO-Network/identity-api/graph/model"
 	"github.com/DIMO-Network/identity-api/internal/config"
 	"github.com/DIMO-Network/identity-api/internal/helpers"
-	"github.com/DIMO-Network/identity-api/internal/repositories"
+	"github.com/DIMO-Network/identity-api/internal/repositories/base"
 	"github.com/DIMO-Network/identity-api/models"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/assert"
@@ -35,7 +35,7 @@ func TestGetManufacturer(t *testing.T) {
 		assert.NoError(t, err)
 	}
 
-	controller := Repository{repositories.New(pdb, config.Settings{})}
+	controller := Repository{base.New(pdb, config.Settings{})}
 	for i := 0; i < 6; i++ {
 		tokenID := i
 		res, err := controller.GetManufacturer(ctx, model.ManufacturerBy{TokenID: &tokenID})
