+++ internal/repositories/reward/rewards_test.go
@@ -10,7 +10,7 @@ import (
 	gmodel "github.com/DIMO-Network/identity-api/graph/model"
 	"github.com/DIMO-Network/identity-api/internal/config"
 	"github.com/DIMO-Network/identity-api/internal/helpers"
-	"github.com/DIMO-Network/identity-api/internal/repositories"
+	"github.com/DIMO-Network/identity-api/internal/repositories/base"
 	"github.com/DIMO-Network/identity-api/models"
 	"github.com/DIMO-Network/shared/db"
 	"github.com/DIMO-Network/shared/dbtypes"
@@ -47,7 +47,7 @@ func (r *RewardsRepoTestSuite) SetupSuite() {
 		DIMORegistryAddr:    "0x4de1bcf2b7e851e31216fc07989caa902a604784",
 		DIMORegistryChainID: 80001,
 	}
-	r.repo = &Repository{repositories.New(r.pdb, r.settings)}
+	r.repo = &Repository{base.New(r.pdb, r.settings)}
 	r.paginationHelper = helpers.PaginationHelper[RewardsCursor]{}
 
 }
