+++ internal/repositories/vehicleprivilege/vehicle_privileges_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/DIMO-Network/identity-api/graph/model"
 	"github.com/DIMO-Network/identity-api/internal/config"
 	"github.com/DIMO-Network/identity-api/internal/helpers"
-	"github.com/DIMO-Network/identity-api/internal/repositories"
+	"github.com/DIMO-Network/identity-api/internal/repositories/base"
 	"github.com/DIMO-Network/identity-api/models"
 	"github.com/DIMO-Network/shared/db"
 	"github.com/stretchr/testify/suite"
@@ -35,7 +35,7 @@ func (s *VehiclesPrivilegesRepoTestSuite) SetupSuite() {
 		DIMORegistryAddr:    "0x4de1bcf2b7e851e31216fc07989caa902a604784",
 		DIMORegistryChainID: 80001,
 	}
-	s.repo = &Repository{repositories.New(s.pdb, s.settings)}
+	s.repo = &Repository{base.New(s.pdb, s.settings)}
 }
 
 // TearDownTest after each test truncate tables
